Finished hw

+++ include/lpm_trie.h
@@ -3,6 +3,7 @@
 // this is inneficient
 // TODO make this trie collapse prefixes
 #include "protocols.h"
+#include "ip.h"
 #include <stdlib.h>
 #include <stdint.h>
 
@@ -18,5 +19,5 @@ typedef struct {
 
 lpm_trie_t *lpm_trie_create();
 void lpm_trie_destroy(lpm_trie_t *trie);
-void lpm_trie_insert(lpm_trie_t *trie, uint32_t prefix, uint32_t mask_len, uint32_t next_hop, int net_interface);
+void lpm_trie_insert(lpm_trie_t *trie, uint32_t prefix, int8_t mask_len, uint32_t next_hop, int net_interface);
 int lpm_trie_lookup(lpm_trie_t *trie, uint32_t ip, uint32_t *next_hop, int *net_interface);
